Added clip store
Now adding clips to local storage.  My intention is to allow you to
gather a bunch of clips at once, then either copy them to clipboard as
a big chunk of markdown or export them (via zip file) as a qvnote, md,
or enex.  I’ll need to implement a store limit.